backend/requirements.txt
pymongo
dotenv
pydantic
pydantic-settings
fastapi
passlib
pydantic[email]
python-multipart
uvicorn
bcrypt
python-jose
google-generativeai
cloudinary
numpy 
librosa 
scipy 
openai-whisper  
fastdtw 
matplotlib 
seaborn 
scikit-learn 
python-Levenshtein 
jupyter 
tqdm
speechbrain
ffmpeg-python
ffmpeg
